read verbosity flag from recProperties
The classes recNorm and recModel also read the verbose flag from the
dictionary recProperties, as recModel has been doing. Up to now, the
boolean verbose_ of recNorm and recPath was initialized to "false" in the
initialisation list of the constructor. This made all if(verbose_)
tests in child classes of recNorm and recPath useless.