recurrence tutorial: laminar vortex shedding
This 2D case of vortex shedding in laminar cross-flow demonstrates the
concept of recurrence.

This is a bit work-in-progress: please check whether this tutorial runs with
the recurrence model and tools of CFDEMcoupling, namely rStatAnalysis.

The simulation roughly goes through three stages:
  * The initial solution computed by potentialFoam
  * A period of symmetric, steady-state flow
  * Finally, periodic vortex shedding

These three stages are clearly visible in the recurrence plot.

  * We see how not one of the later velocity fields is similar to the initial one
  * We see the intermediate stage with a symmetric flow field
  * We see the periodic vortex shedding