Recurrence-based turbulence: model and test-app
A base class for recurrence-based turbulence models was created. This class
holds a reference to the recurrenceModel in use.
Thus, turbulent fields can be updated from the data base.

The recurrence-based turbulence models are essentially re-implementations
of the respective standard turbulence models. In addition to being derived
from their respective base class, as mandated by OpenFOAM, each recurrence-based
turbulence model is derived from the class recurrenceTurbulenceModel.
This allows for making use of polymorphism on the recurrence-part of the
turbulence model, as after construction, the solver needs to pass the reference
to the recurrenceModel to the recurrence-based turbulence model.