communicate particle cell ids to liggghts and back.
this allows for use of a turboEngine-like locate section
with an initial guess for the cell.

to use this, add a
`fix             prev_cell_ids all property/atom prev_cell_ids scalar no
no no -1`
to the liggghts run script and set `useCellIdComm yes` in the
twoWayOne2OneProps section of the couplingProperties.

a check whether the particle is contained in the bounding box
of the domain is added regardless of the setting.